backend/DAM-ProjectCore/requirements.txt
falcon~=3.0.0
passlib~=1.7.4
sqlalchemy==1.3.23
sqlalchemy_i18n
gunicorn
pymysql~=1.0.2
jsonschema
falcon-multipart
Jinja2~=2.11.3
kwargs~=1.0.1
dateutils